// tsundoku/blueprints/ux/static/ts/api/types.gen.ts
export type Library = {
  /**
   * Id
   */
  id_: number;
  /**
   * Folder
   */
  folder: string;
  /**
   * Is Default
   */
  is_default: boolean;
};